Added for loops
Something is wrong with my blocks though. If it doesn't end with
and expression it won't compile properly. It seems to work as long
as that is the case though. For example:

for (var i = 0; i < 10; i = i + 1) {
    print(i);
    i + 1;
}

If the `i + 1;` is omitted though it complains that it expected
an expression.